feat: 634 Library - Use LocalStorage (#647)
* Modify Create Lib
* Modify Delete library Lib
* Modify Update Lib

+++ Application/Libraries/Delete/DeleteLibraryCommandHandler.cs
@@ -4,9 +4,10 @@ using Domain.Primitives;
 using MediatR;
 using Application.Interfaces;
 using MongoDB.Bson;
+using Persistence.LocalStorage;
 
 namespace Application.Libraries.Delete;
-internal sealed class DeleteLibraryCommandHandler(IRepository<Library, ObjectId> librayRepository, IUnitOfWork unitOfWork) : IRequestHandler<DeleteLibraryCommand, Result>
+internal sealed class DeleteLibraryCommandHandler(IRepository<Library, ObjectId> librayRepository, IUnitOfWork unitOfWork, ILibraryLocalStorage libraryLocalStorage) : IRequestHandler<DeleteLibraryCommand, Result>
 {
     public async Task<Result> Handle(DeleteLibraryCommand request, CancellationToken cancellationToken)
     {
@@ -15,6 +16,12 @@ internal sealed class DeleteLibraryCommandHandler(IRepository<Library, ObjectId>
         if (library is null)
         {
             return LibrariesError.NotFound;
+        }
+
+        var result = libraryLocalStorage.Delete(library.RelativePath);
+        if (result.IsFailure)
+        {
+            return LibrariesError.FolderNotDeleted;
         }
 
         librayRepository.Remove(library);
